+++ changelog.md
@@ -14,6 +14,12 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 
 # v1.8.1
 
+## Changed
+
+* Updated polkadot-sdk dependency to polkadot-stable2509.
+Partner Chains Aura modification follows changes regarding checking inherents that are also present in the new polkadot-sdk release.
+Pallet Session new parameters should be: `type KeyDeposit = ();` and `type Currency = Balances;`.
+
 ## Fixed
 
 * Fixed offchain error when irrelevant utxos were present while updating d-param and permissioned candidates